Enable context menu rename/move/delete of album collection views
